requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.0
